split UT files
